添加尸体

+++ src/YT.Application/Cards/Dtos/CardEditDto.cs
@@ -27,7 +27,7 @@ namespace YT.Cards.Dtos
         /// </summary>
         [DisplayName("金额")]
         [Required]
-        public decimal Money { get; set; }
+        public int Money { get; set; }
 
         /// <summary>
         /// 是否已使用
